Add partial support for wl_drm and linux-dmabuf protocols
This commit adds support for DMABUF file descriptor types and for the
wayland-drm and linux-dmabuf-v1 protocols, on machines with an i915
driver only. (Other drivers are not yet supported, due to difficulties
with the generic libgbm buffer management library.)

To support the wl_drm protocol, the device advertisement event was modified
to indicate only a default DRM render node (/dev/dri/renderD128). This
avoids the authentication steps which would otherwise be necessary when
the main DRM device paths are provided, and still should work for most
OpenGL programs.

This change also involved significant main loop refactoring to remain
readable during an extensions to the protocol parsing code; specifically,
to allow message handlers to increase the length of protocol messages.

A command line flag, -no-gpu, has been added to block the wl_drm and
zwp_linux_dmabuf_v1 global advertisements; when running on a system
without an i915 drm node, these two protocols are also hidden, to prevent
applications from providing waypipe with buffers that it cannot handle.

The program still compiles when libdrm_intel is not available.

Because tiled buffers do not appear to work between systems with
gpus of different generations, nonstandard dmabuf format+modifier
advertisements are now filtered out.