Switch dmabuf backend to libgbm
The previous intel-specific, libdrm-based backend for dmabuf
operations was a temporary measure. This commit replaces
it with a backend based on libgbm, a generic buffer management
library supported by most graphics drivers.

DMABUFs are still assumed to correspond to standard single-plane
linear packed image. This limitation can be corrected, but was
necessary for this commit because buffers are currently imported
and created before protocol messages are read.