Use vectored io to reduce data copies when writing to channel
Specifically, the function pack_pipe_message, instead of allocating
a large buffer and copying transfer blocks into the buffer, now sets
up metadata for the function writev, which can writev a series of
data buffers without unnecessary copies.

This is especially useful for file or dmabuf update diffs, which are
often >100KB; it reduces the net processing time for them by 20%, in
some cases.