Deduplicate dmabufs for multiplanar images
A common construct when creating multiplanar dmabufs for the linux-dmabuf
protocol is to allocate all planes as part of the same dmabuf, and then
to provide the same file descriptor to the protocol for each individual
plane. As this file descriptors is dup2'd on being sent from an application
to waypipe, determining which file descriptors share an underlying
dmabuf requires extracting their associated buffer handles with libgbm.

Additional error checks for close() operations, and to verify the validity
of file descriptors before sendmsg'ing to another program, have also been
added.