Clarify reference counting logic for struct shadow_fd
This commit divides the reference count for struct shadow_fd objects
into two main categories: the reference counts required to keep
the file descriptors associated with the shadow_fd alive until they
can be written to the Wayland application/compositor, and the reference
counts from protocol objects.

In the process, fix a memory leak and an early destruction bug for dmabuf
objects.