Track shm buffer damage using collections of rectangles
This commit replaces the existing interval-based rectangle damage
tracking (which selected the largest interval containing all damaged
rectangles) with a more fine grained mechanism to replace the set of
damaged rectangles with a set of disjoint rectangles that covers
it.

Because the damage buffer diff procedure is optimized for linear
scans, damage rectangles are actually represented in terms of
the damage they apply to the underlying (linear layout) image buffer.
These "extended intervals" consist of a series of uniformly spaced
intervals, and permit a new (but minor) optimization: rectangles along
the left and right borders of an image can be merged, since their
associated data is closely placed in memory.

The rectangle damage tracking routine here both uses a naive O(n^2)
algorithm and does not yet minimally combine two overlapping extended
intervals, instead producing the simple interval containing the two.