More finegrained damage merging, bugfixes
In addition to corrections for a number of small errors, this change
ensures that e.g. nearby parallel vertical lines are merged into the
smallest rectangle which contains both. Detecting the case where
the two lines are more than the merging margin away from each other
in memory has not yet been done.