Speed up damage merge algorithm
By processing the damage records in sorted order, the updated damage
merge algorithm ensures that damage rectangles which no longer
can intersect with any unmerged rectangles can be removed from the pool
of known-disjoint damage rectangles early.

The updated interval is faster most of the time, but can be slightly slower
sometimes if the set of damage rectangles would merge more efficiently
in unsorted order, or when rectangles make many passes to try to enter
the disjoint pool before succeeding; the binary heap is not optimized
for 'hot' data that is reused shortly after insertion. (A fibonacci heap
may be a faster alternative.)