Add support for additional pixel video formats
Not that many formats are supported, as the ffmpeg pixel formats
do not always coincide with DRM formats, because h264 does support
transparency, and for lack of test programs that use any of the
other formats.

This change also changes the signal used to notify when video
encoding is used; it is now a flag in the special field of the
data transfer header.