Remove redundant check for buffer changes
Using memcmp to pre-scan a buffer for changes only saves time in
the event that there are no changes whatsoever, as otherwise
the diff routine must still be run on the entire damaged buffer
region to identify the change. Since the dirty flags maintained
by the protocol are generally accurate, removing this initial
optimistic scan does not significantly add to runtime.

The now-removed memcmp did slow down diff construction for e.g.
terminals that don't provide damage tracking, and tend to have
many small changes clustered near the prompt at the base of
the terminal.