Partition diff construction evenly over damage area
As opposed to partitioning the diff evenly over the entire buffer area,
which, when damage is concentrated e.g. at the top or bottom of an image,
takes longer than when each thread handles an even fraction of the damage
area. (While this can lead to a tiny bit more cache migration for the
buffer, 1] that already happens when the program updates 2] damage request
patterns tend to be consistent over time anyway.)