Add fuzzer-friendly test program to run main loop
The new test program runs, in separate threads, a local/display-side
main loop, a remote/application-side main loop, and a "controller"
script that reads messages from an input file and sends them
to the main loops. (Each pair of threads is linked via socketpair.)

The input file format consists of a series of headers followed by data.
The headers indicate the length of the following data, and to which of
the application or compositor the message should be sent. At the moment,
sending file descriptors is not supported.