Fix a few protocol parsing/handler crashes
It was previously possible to cause a use-after-free when creating new
objects by setting the new object id to match that for the method being
called.

A very similar bug in which object replacement broke invariants with
non-zombie targets was fixed.

This change also adjusts fuzz_hook to use a more thread-safe log handler,
gated behind a command line option to reduce the number of context
switches.