Add support for hardware video decoding with VAAPI
This commit adds a --hwvideo option to waypipe, which enables
video encoding of dmabufs and ensures that waypipe will use VAAPI
to encode video, if available and if formats permit. This mode
can be faster than libx264, although it typically requires a
larger bitrate for the same quality level.

This new option is still rather unstable, and crashes with some
programs. It also has restrictions on buffer sizes; windows with
fixed sizes divisible by e.g. 32 are preferred.