Add an actual decoding pathway using VAAPI + sws_scale
When --hwvideo is enabled, this commit uses VAAPI to decode if possible.
For lack of an easy way to perform NV12 -> RGB format conversion on the
GPU using ffmpeg, any resulting frames are copied back to the CPU,
converted via libswscale, and then written to the DMABUFs on the GPU
again.