Adjust video buffer alignment, and disabled unaligned hwvideo
Currently, using VAAPI to encode and decode fails when the buffer
stride is not divisible by 64; but instead of failing cleanly,
the underlying dmabuf appears to be modified so that trying to
map it to CPU memory fails. This commit preemptively disables
hardware video decoding for such unaligned buffers.

It also ensures that when hardware encoding is used that the encode
and decode buffer sizes match.