Make channel fd replaceable while program is running
The existing channel connection can now be replaced through a new
connection specified by writing to the server control pipe.

Because the connection closing no longer definitely implies that the
remote program is inaccessible, a new specialized WMSG_CLOSE message
type has been added to signal this.

The low bit of the initial (re)connection tokens is used to let the
display-side waypipe reject attempts to reconnect with a subprocess
that no longer exists, rather than creating a new process with that
token.