Add SSE, AVX, and NEON optimized diff routines
The conditional compilation of the different routines is handled using
meson's unstable simd module.

The vectorized diff routines themselves are relatively straightforward
translations of the plain C diff routine. Additional alignment constraints
are imposed to ensure that loads and stores to the mapped buffer and its
mirror are aligned; in some cases this can help performance.

The output diff format has not been changed beyond a small adjustment
to the handling of trailing data; while it would have been possible to
use e.g. distinct diff data and control buffers, the performance gain
for such a change is negligble.