Revert to simpler LZ4 and Zstd APIs
As adding frame headers to compressed data is no longer needed, it
is safe to revert from the LZ4F API to the simpler LZ4 and LZ4HC APIs
for fast and slow LZ4 compression modes.

Similarly, as fine grained control of the Zstd compression parameters
is neither needed nor used, the use of the very new ZSTD_compress2
function is replaced with the much older and equivalent ZSTD_compressCCtx
function.