Fix building NEON routines with Clang on ARMv6
Clang only compiles NEON code for armv(>=7). To ensure that it is still
possible to compile waypipe to run on armv6, and run on armv7 with neon,
the compile architecture is increased only for the code that uses neon.

Because architectures are not forward compatible, the code to check at
runtime which instruction sets can be used has been moved to kernel.c,
where the increased architecture version can not affect it.