Simplify locking for transfer_data
... by splitting struct transfer_data into a struct transfer_queue, which
needs to be locked to work correctly, and a struct chan_write_queue,
which the mainloop uses as an input to writev, and which does not have
any locks.

As writev sometimes takes 0.1msec, and used to be performed under
the transfer_data lock, this change reduces lock acquisition times
significantly for worker threads that try to add new messages to be
transferred.