Use DEFAULT_SOURCE except for the gnu platform.c
DEFAULT_SOURCE was chosen in favor of POSIX_C_SOURCE 200809L
because the former includes CMSG_LEN/CMSG_SPACE, which are needed
to portably align cmsg entries.

The OS-specific code has been moved to platform.c.

As IOV_MAX is an upper bound (and at runtime, the value is allowed
to be different), it is now determined using sysconf().

The pipe_mirror test has been adjusted to allow extra transfer modes
when pipes are bidirectional.