Post Wayland error event on fatal error
The main loop code (and some of shadow.c) has been adjusted to
use different error codes for allocation failure and for regular
errors.

If there is an error, then the connection closing message will
indicate the type (oom/error/natural connection close). For the
first two types, a corresponding error event will also be sent to the
Wayland application.