Add -Wno-unused-result compile flag
On some systems with gcc, a (void)cast does not suffice to ignore
the return values of read() and write(). However, in some cases,
such as when closing Waypipe via SIGINT, the program will still
shut down no matter the return value of the write() used to indicate
that a SIGINT occured.