Annotate Wayland messages over wire with number of fds
Tag each Wayland message sent over the wire with the number of file descriptors
that Waypipe's logic has associated with the message. This tag makes it easier
to detect if the Waypipe client and server instances, for some reason, have
interpreted the two streams of file descriptors differently. (For example, this
may be useful if one instance supports a new Wayland protocol involving a file
descriptor, but the other does not.)

This change introduces a limit of 31 fds per method and 2047 different methods
per Wayland protocol type. It is very unlikely that a well written protocol will
ever surpass either of these two limits.