Treat POLLHUP like POLLIN and check for EOF
It is possible for poll to indicate POLLHUP and still have outstanding
data to be read. To ensure that this data is read, the main loop code
now treats POLLHUP as POLLIN, and only acts on the hang-up when the
socket being read from actually returns EOF/zero bytes read.