Fix connection endianness detection
Change location of test bit for endianness detection to something that is
guaranteed not to be set if the byte order is wrong.

When the bit to check endianness is not set, there are two possibilities:
a) the endianness of the remote system actually differs
b) a generally bad connection token, e.g. from older Waypipe versions