Fix buffer damage tracking with multiple buffers
The damage requests for the wl_surface interface only indicate the
region of the surface which has changed since the last commit. This
commit computes the net change to a given buffer used by a surface
by accumulating all the surface damage since the last time the buffer
was presented.