Only update shm_pool fds when surfaces damaged
This change use the fact that calloc and ftruncate provide
zero-initialized memory effectively for free, so it is easy to
create consistent initial states for files + mirror buffers. It also
makes Waypipe more efficient with applications that allocate huge shm
pools but only use a tiny fraction of them.

This change may slightly slow down buffer creation in the usual case,
because the first update will use the 'diff' instead of the 'fill'
command.