Check size of buffer when mapping dmabuf
For context, a DMABUF which was created with a given stride need not
be mapped with that stride, as the original format may be nontrivial,
while gbm_bo_map always provides a linear layout. This is a slight
workaround for the issue that the strides that DMABUFS are mapped with
may differ between systems -- to give up if the destination is too small.

This is a minor bandage over the real issue, which is that Waypipe
currently treats DMABUFS as buffers (for which the stride and layout
is critical), while gbm treats them more like images (where the
stride and layout don't really matter). Fixing this properly may
require wire protocol changes in the future.