Implement threadless fuzz hook
Adapting the framework from the protocol control test, we can
create a fuzz hook which does not require running the Waypipe
main loop in a separate thread, and is thus fully deterministic.
The tradeoff is that this fuzz hook cannot test details (like
reconnection machinery) that are only available in the main loop.