Delay switch to default LZ4 compression
While using waypipe ssh to connect the development version of Waypipe
to v0.8.1 works despite the change in default compression, because
waypipe ssh now always sets the remote compression type, the reverse
does not hold; connecting v0.8.1 to the devel version with waypipe ssh
fails because v0.8.1 does not set '-c none' in the remote.

This commit reverts the default compression type of Waypipe back to
'none', to avoid this incompatibility. It will be switched to 'lz4'
after the next release.