Rewrite socket connect/bind/unlink handling
This commit changes the way socket files are bound, connected to, and
unlinked. Instead of directly operating on the socket path, Waypipe now
opens the folder which will contain the socket, fchdirs to it, operates
on the socket filename, and then fchdirs back to the current working
directory.

This new approach:
* Allows for opening slightly longer file paths, as long as the filename
  part doesn't exceed sizeof(sockaddr_un->sun_path).
* Avoids some race conditions where the folder part of the path is renamed
  and replaced by a new folder between the times bind and unlink are called.

Libwayland does not support longer file paths, so this change does not
have any effect for typical programs.

Because Waypipe is not a library, and the relevant code is single threaded;
calling fchdir to go back and forth between folders does not interfere
with any other operations.

If POSIX ever gains support for connectat, bindat, and unlinkat, then
the fchdir trick can be removed.